+++ PSPDFKit.embeddedframework/PSPDFKit.framework/Versions/A/Headers/PSPDFPageRenderer.h
@@ -27,6 +27,11 @@ extern NSString *const PSPDFRenderAllowAntiAliasing;       // Enabled/Disables a
 extern NSString *const PSPDFRenderBackgroundFillColor;     // Allows custom render color. Default is white.
 extern NSString *const PSPDFRenderPDFBox;                  // Allows custom PDF box (if pageInfo is nil)
 
+// Allow custom content rendering after the PDF. Type is `PSPDFRenderDrawBlock`.
+extern NSString *const PSPDFRenderDrawBlockKey;
+
+typedef void (^PSPDFRenderDrawBlock)(CGContextRef context, NSUInteger page, CGRect cropBox, NSUInteger rotation, NSDictionary *options);
+
 // Can be used to use a custom subclass of the `PSPDFPageRenderer`. Defaults to nil, which will use `PSPDFPageRenderer.class`.
 // Set very early (in your AppDelegate) before you access PSPDFKit. Will be used to create the singleton.
 extern Class PSPDFPageRendererClass;
@@ -38,14 +43,14 @@ extern Class PSPDFPageRendererClass;
 + (instancetype)sharedPageRenderer;
 
 /// Setup the graphics context to the current PDF.
-- (void)setupGraphicsContext:(CGContextRef)context inRectangle:(CGRect)displayRectangle pageInfo:(PSPDFPageInfo *)pageInfo;
+- (void)setupGraphicsContext:(CGContextRef)context rectangle:(CGRect)displayRectangle pageInfo:(PSPDFPageInfo *)pageInfo;
 
 /// Renders a page inside a rectangle. Set context CTM and clipRect to control rendering.
 /// Returns the renderingRectangle.
-- (CGRect)renderPageRef:(CGPDFPageRef)page inContext:(CGContextRef)context inRectangle:(CGRect)rectangle pageInfo:(PSPDFPageInfo *)pageInfo withAnnotations:(NSArray *)annotations options:(NSDictionary *)options;
+- (CGRect)renderPageRef:(CGPDFPageRef)page inContext:(CGContextRef)context rectangle:(CGRect)rectangle pageInfo:(PSPDFPageInfo *)pageInfo annotations:(NSArray *)annotations options:(NSDictionary *)options;
 
 /// Renders a page; defined by point and zoom. Use `zoom=100` and `point = CGPointMake(0.f, 0.f)` for defaults.
-- (CGSize)renderPage:(CGPDFPageRef)page inContext:(CGContextRef)context atPoint:(CGPoint)point withZoom:(double)zoom pageInfo:(PSPDFPageInfo *)pageInfo withAnnotations:(NSArray *)annotations options:(NSDictionary *)options;
+- (CGSize)renderPageRef:(CGPDFPageRef)page inContext:(CGContextRef)context atPoint:(CGPoint)point withZoom:(double)zoom pageInfo:(PSPDFPageInfo *)pageInfo annotations:(NSArray *)annotations options:(NSDictionary *)options;
 
 /// Renders a particular appearance stream (A PDF within a PDF) into a context.
 /// Will return NO if rendering failed.